+++ src/Features/Sitecore.Pathfinder.PageHtml/PageHtml/PageHtmlGetRendererProcessor.cs
@@ -1,7 +1,7 @@
+// © 2015-2016 Sitecore Corporation A/S. All rights reserved.
+
 using System;
-using Sitecore.Mvc.Extensions;
 using Sitecore.Mvc.Names;
-using Sitecore.Mvc.Presentation;
 using Sitecore.Pathfinder.Extensibility.Pipelines;
 using Sitecore.Pathfinder.Mvc.Response.GetRenderer;
 
@@ -12,12 +12,6 @@ namespace Sitecore.Pathfinder.PageHtml.PageHtml
         // must come before HtmlGetRendererProcessor
         public PageHtmlGetRendererProcessor() : base(500)
         {
-        }
-
-        [Diagnostics.CanBeNull]
-        private string GetViewPath([Diagnostics.NotNull] GetRendererPipeline pipeline)
-        {
-            return GetViewPathFromRenderingType(pipeline) ?? GetViewPathFromRenderingItem(pipeline);
         }
 
         protected override void Process(GetRendererPipeline pipeline)
@@ -35,36 +29,20 @@ namespace Sitecore.Pathfinder.PageHtml.PageHtml
 
             pipeline.Renderer = new PageHtmlRenderer
             {
-                FilePath = filePath,
-                Rendering = pipeline.Rendering
+                FilePath = filePath
             };
 
             pipeline.Abort();
         }
 
         [Diagnostics.CanBeNull]
-        protected virtual string GetViewPathFromInnerItem([Diagnostics.NotNull] Rendering rendering)
+        private string GetViewPath([Diagnostics.NotNull] GetRendererPipeline pipeline)
         {
-            var filePath = rendering.RenderingItem.InnerItem["Path"];
-            return string.IsNullOrWhiteSpace(filePath) ? null : filePath;
+            return GetViewPathFromRenderingType(pipeline) ?? GetViewPathFromRenderingItem(pipeline);
         }
 
         [Diagnostics.CanBeNull]
-        protected virtual string GetViewPathFromLayoutItem([Diagnostics.NotNull] GetRendererPipeline pipeline)
-        {
-            var filePath = pipeline.LayoutItem.ValueOrDefault(item => item.FilePath);
-            return string.IsNullOrWhiteSpace(filePath) ? null : filePath;
-        }
-
-        [Diagnostics.CanBeNull]
-        protected virtual string GetViewPathFromPathProperty([Diagnostics.NotNull] Rendering rendering)
-        {
-            var filePath = rendering["Path"];
-            return string.IsNullOrWhiteSpace(filePath) ? null : filePath;
-        }
-
-        [Diagnostics.CanBeNull]
-        protected virtual string GetViewPathFromRenderingItem([Diagnostics.NotNull] GetRendererPipeline pipeline)
+        private string GetViewPathFromRenderingItem([Diagnostics.NotNull] GetRendererPipeline pipeline)
         {
             var renderingItem = pipeline.Rendering.RenderingItem;
             if (renderingItem == null || renderingItem.InnerItem.TemplateID != TemplateIds.ViewRendering)
@@ -77,21 +55,24 @@ namespace Sitecore.Pathfinder.PageHtml.PageHtml
         }
 
         [Diagnostics.CanBeNull]
-        protected virtual string GetViewPathFromRenderingType([Diagnostics.NotNull] GetRendererPipeline pipeline)
+        private string GetViewPathFromRenderingType([Diagnostics.NotNull] GetRendererPipeline pipeline)
         {
             if (pipeline.Rendering.RenderingType == "r")
             {
-                return GetViewPathFromInnerItem(pipeline.Rendering);
+                var filePath = pipeline.Rendering.RenderingItem.InnerItem["Path"];
+                return string.IsNullOrWhiteSpace(filePath) ? null : filePath;
             }
 
             if (pipeline.Rendering.RenderingType == "View")
             {
-                return GetViewPathFromPathProperty(pipeline.Rendering);
+                var filePath = pipeline.Rendering["Path"];
+                return string.IsNullOrWhiteSpace(filePath) ? null : filePath;
             }
 
             if (pipeline.Rendering.RenderingType == "Layout")
             {
-                return GetViewPathFromLayoutItem(pipeline);
+                var filePath = pipeline.LayoutItem.FilePath;
+                return string.IsNullOrWhiteSpace(filePath) ? null : filePath;
             }
 
             return null;