+++ src/Sitecore.Pathfinder.Core/Packaging/IPackageService.cs
@@ -13,11 +13,13 @@ namespace Sitecore.Pathfinder.Packaging
         [NotNull, ItemNotNull]
         IEnumerable<IPackage> CheckForInstalledUpdates([NotNull, ItemNotNull] IEnumerable<IPackage> installedPackages);
 
-        [NotNull, ItemNotNull]
-        IEnumerable<IPackage> FindInstalledPackagesById([NotNull] string packageId);
+        void DownloadPackage([NotNull] string packageId, [NotNull] string version, [NotNull] string fileName);
 
         [NotNull, ItemNotNull]
         IEnumerable<IPackage> FindInstallablePackagesById([NotNull] string packageId);
+
+        [NotNull, ItemNotNull]
+        IEnumerable<IPackage> FindInstalledPackagesById([NotNull] string packageId);
 
         [NotNull, ItemNotNull]
         IEnumerable<IPackage> GetInstallablePackages([NotNull] string queryText, [NotNull] string author, [NotNull] string tags, int skip = -1);