+++ Changes.md
@@ -18,6 +18,8 @@ Version Next
 * Add: Renaming and deletes now works by stamping the Project Unique ID into the __PathfinderUniqueIds fields on every item on the server (2016-04-16)
 * Add: Merged 'Layout.HtmlFile' and 'Layout.Page' into single feature 'Layout.File' (2016-04-16)
 * Add: show-config task (2016-04-19)
+* Add: show-website task (2016-04-20)
+* Add: restore-packages task for downloading dependency Nuget package (2016-04-20) - Dmitry Kostenko
 
 Version 0.6.0
 =============